requirements.txt
flask==2.0.1
flask-cors==3.0.10
numpy==1.21.0
opencv-python==4.5.3.56
networkx==2.6.3
pillow==9.5.0
rasterio==1.2.10
python-dotenv==0.19.0
torch==1.9.1
torchvision==0.10.1
scikit-image==0.19.0
segmentation-models-pytorch==0.3.0 